requirements.txt
rich
fade
bleak
python-dotenv
